requirements.txt
requests
trafilatura
python-dotenv
tldextract
groq
discord.py
google-cloud-bigquery
google-cloud-secret-manager
lxml_html_clean
asyncio
